give a context to all V.one calls
When One objects also have a configuration space, we can track that they are used correctly, i.e., that we don't use Ones in a context that is broader than in which they were created.
This is potentially slower than necessary and could be removed in a final version, but this is useful for checking invariants and finding bugs.

Additionally a map is introduced for lambda methods to enable reusing methods that have the same name (used for PUT only right now)