track variable names and refactor LiftUtils
variables and parameters now track names and types, for debugging purposes (can be kept for an IDE in the .class file, but that part is not yet implemented)

additionally changed LiftUtils from a trait to an object. Those were all global static methods anyway, so letting various classes inherit that trade didn't make a lot of sense. Importing those global functions now instead.