create variable names in bytecode for debugging
storing the names and types of local variables (including parameters)
in the .class file. Now debuggers can show names for the local variables.

currently simply indicating that all variables are valid throughout
the entire source code. we might want to invest into an analysis
to get the precise scope, but this doesn't seem very important.

not sure whether we can find a better way to name the block context
variables