fix complications with vblock
the last vblock may not necessary start at the last position, so that there is a danger of not jumping back. now all blocks of the last vblock are serialized last.
we might consider reodering blocks by vblock in general to avoid confusion, but we need to make sure that the fallthrough edges remain equivalent (if necessary need to introduce goto statements).
this is not an issue on the last vblock though, as it only will be entered once all other vblocks have unsatisfiable conditions.