processing of atomic exceptions in VBlocks
infrastructure to catch atomic exception (with several test cases for division by zero exceptions)

redesign of the exception handling mechanism. Instead of separate VBlocks with their own variables,
exception handlers are always part of the same VBlock. If multiple blocks in different VBlocks share
an exception handler, we introduce new exception handler blocks that just jump to the original block,
however those new blocks can be in the same VBlock as the block throwing the exception and the
rest is handled by the control-flow mechanism.