Able to run GPL
However, seems to have a bug in VLinkedList, but executing in debug mode
works fine. TestLaunch of GPL is not working yet because trace
comparison failed.

The way we call library methods is again quite different from before.
Now we use @Immutable tag to distinguish library classes. Current form
is far from perfect, but seems to work well. We need to write a bunch of
model classes, but the process could be automated.

A bunch of small changes and bug fixes:
* For classes that implement Java interfaces, we create wrapper method
to call our variational method, because the call site of that method is
usually inside JDK.
* Initialized all non-conditional fields to be One(null), instead of
 null, otherwise we need null checker everywhere (e.g. VOps)
* ... (trying to remember)