Run Email.
Fix some bugs in method renaming. Forgot to rename some return types
to their corresponding model classes.

Modify one assertion in VBlockAnalysis, which checks either all or none of the successors should be in the current VBlock. If the current
VBlock only contains one Block, this check could be avoided.