Fix class loading problem.
Antlr uses Thread.getContextClassLoader() to load some class (DetailAST
in this case), which has already been loaded by our VBCClassLoader
before. A ClassCastException was thrown because the same class is loaded
 by two different class loaders.

To fix this, we set the contextClassLoader of the current thread to be
our VBCClassLoader, in Launcher.scala.

Another problem is VBCClassLoader tries to load the same class again and
 again, which technically create different Class objects. To fix this,
 we do caching every time we load a new class.