Fix a whenEQ() issue
In rare cases, it might happen that the same local variable is used to
store values of different types. If one of them is boolean (or int) and
it is passed to whenEQ(), it is likely that the underlying when() call
would perform predicate test on some strange object (e.g., DetailAST).

To avoid that, we filter out values with unsatisfiable context before
performing predicate test.