Add is64Bit field to V_TYPE
This new field is used to tell if a V is representing a 64-bit value or
not. Providing a default value to this field would save a lot of
refactoring, but it is also dangerous because each instruction needs to
be considered carefully. This refactoring enables us to use our DFA to
do a more fine-grained analysis, but it also complicates the rewriting
process of return instructions because exceptions are treated as return
values. For this reason, we couldn't tell the exact V_TYPE for local
variables because both kinds of V_TYPE could be stored at the same time
(e.g., one exception and one long value).

For completeness of refactoring, we also add the FSTORE instruction.

Because of this refactoring, instructions like POP2 and DUP2 can be
implemented easily.