Provide a few unlifted methods
In some rare cares, unlifted JDK code would call some program methods
that implement special superclasses or interfaces (e.g. run() of
Runnable). Since we are lifting the original methods, we need to provide
 a fake entry points (similar to main methods). Usually this operation
is risky because we do not know the accurate context, unless we have
some ways to track context of each thread.

For now, we assume that all threads are created under context True, and
this assumption seems to be reasonable in Jetty.