simple implementation to check correct vblock order
as discussed, vblock-order should follow the post-dominator
relationship, otherwise joins may be inefficient.

right now only checks are implemented and since no violation
was found yet, no need to reorder blocks yet.

in fact, there are a number of violations for simple loops, where
the body of the loop is dominated and postdominated by the
block with the loop header. if a node is both dominated and
postdominated by the same other node, the order does not matter
and is not enforced.