attempted to rewrite infrastructure for providing configurations during tests.
unfortunately, I may have broken more than fixed in the current version,
because many assumptions are broken:

If we want to lift bytecode generally (and not separately for every
configuration), we should provide configurations as a parameter,
not as some magic bytecode rewrite.

If we want to run multiple configurations in the same loaded classes
we cannot rely on static fields, which are only initialized once for
multiple runs.

This may require a more reasonable design of the test subjects in which
we pass around configuration parameters instead of using global
static fields. We could do something like this automatically by
introducing yet another parameter to each method (or use a call graph
analysis to figure out which methods), but that seems fairly wasteful.
The better solution is not to have any magic annotations in the program,
but rather invoke a main method with plain or conditional parameters,
depending on what we want to test, given that the subject systems
have a reasonable configuration mechanism in the first place.