fix bug: variable may be used both as 32 and 64 bit, use wider version
a variable slot may be used for different types. if it's used
both as a 32 and a 64 bit type, we can run into problems
if it's first used as 32 bit and we assume that the variable
is 32 bit since, even though it's later used as a 64 bit.

this caused problems in the constructor of java/awt/GradientPaintContext

now it overwrites the definition of variables with a wider one
when found.