starting to replace the lifting mechanism (incomplete)
all lifted classes are loaded under a new package name
(with a prefix). all classes are rewritten to use these new
lifted classes.

a lifting policy can be used to decide which classes not to lift
(currently only Object and String; String only because it's
used as part of field initializers)

also significantly rewritten to avoid global objects for lifting.

most of the V-translation for method calls is commented out right
now.

goal was to rewrite and rename everything first without actual lifting.